requirements.txt
# Django Framework
Django==4.2.7

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter (for production)

# Image Processing
Pillow==10.1.0

# Face Recognition (Optional - requires additional system dependencies)
# Uncomment if you want real face recognition
# face-recognition==1.3.0
# dlib==19.24.2

# OpenCV for face detection
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# NumPy for numerical operations
numpy==1.26.2

# Environment Variables
python-decouple==3.8

# Excel/CSV Export
openpyxl==3.1.2
xlsxwriter==3.1.9
pandas==2.1.3

# PDF Generation
reportlab==4.0.7
PyPDF2==3.0.1

# Email
django-mail-templated==2.6.5

# Security
django-cors-headers==4.3.1
django-csp==3.8

# API (if needed in future)
djangorestframework==3.14.0

# Caching
django-redis==5.4.0
redis==5.0.1

# Task Queue (for background jobs)
celery==5.3.4
django-celery-beat==2.5.0

# Timezone
pytz==2023.3.post1

# Forms
django-crispy-forms==2.1
crispy-bootstrap5

# Admin Enhancements
django-import-export==3.3.3

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
sentry-sdk==1.38.0

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
faker==20.1.0

# Code Quality
flake8==6.1.0
black==23.11.0
pylint==3.0.3
isort==5.12.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
